CSL: fixes so that g++ is willing to compile some of the code on a 32-bit Linux (eg Raspberry Pi) where it had been wanting to mean about some ambiguities between int64_t and int128_t where 64-bit platforms seemed not to have trouble. So I have put in some explicit casts in a few places so it now builds OK
git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@6527 2bfe0521-f11c-4a00-b80e-6202646ff360

+++ csl/cslbase/arith03.cpp
@@ -453,17 +453,20 @@ static LispObject quotbi(LispObject a, LispObject b)
     int128_t p;
     switch (lena)
     {   case 0: // should never arise but put code in for frivolity!
-            p = bignum_digits64(a, 0);
+            p = static_cast<uint128_t>(bignum_digits64(a, 0));
             break;
         case 1:
-            p = ASL(bignum_digits64(a, 1), 31) | bignum_digits(a)[0];
+            p = static_cast<uint128_t>(
+                    ASL(bignum_digits64(a, 1), 31) | bignum_digits(a)[0]);
             break;
         case 2:
-            p = ASL(bignum_digits64(a, 2), 31) | bignum_digits(a)[1];
+            p = static_cast<uint128_t>(
+                    ASL(bignum_digits64(a, 2), 31) | bignum_digits(a)[1]);
             p = (uint128(p) << 31) | bignum_digits(a)[0];
             break;
         case 3:
-            p = ASL(bignum_digits64(a, 3), 31) | bignum_digits(a)[2];
+            p = static_cast<uint128_t>(
+                    ASL(bignum_digits64(a, 3), 31) | bignum_digits(a)[2]);
             p = (uint128(p) << 31) | bignum_digits(a)[1];
             p = (uint128(p) << 31) | bignum_digits(a)[0];
             break;
@@ -500,17 +503,20 @@ static LispObject quotrembi(LispObject a, LispObject b)
     int128_t p;
     switch (lena)
     {   case 0: // should never arise but put code in for frivolity!
-            p = bignum_digits64(a, 0);
+            p = static_cast<uint128_t>(bignum_digits64(a, 0));
             break;
         case 1:
-            p = ASL(bignum_digits64(a, 1), 31) | bignum_digits(a)[0];
+            p = static_cast<uint128_t>(
+                    ASL(bignum_digits64(a, 1), 31) | bignum_digits(a)[0]);
             break;
         case 2:
-            p = ASL(bignum_digits64(a, 2), 31) | bignum_digits(a)[1];
+            p = static_cast<uint128_t>(
+                    ASL(bignum_digits64(a, 2), 31) | bignum_digits(a)[1]);
             p = (uint128(p) << 31) | bignum_digits(a)[0];
             break;
         case 3:
-            p = ASL(bignum_digits64(a, 3), 31) | bignum_digits(a)[2];
+            p = static_cast<uint128_t>(
+                    ASL(bignum_digits64(a, 3), 31) | bignum_digits(a)[2]);
             p = (uint128(p) << 31) | bignum_digits(a)[1];
             p = (uint128(p) << 31) | bignum_digits(a)[0];
             break;
